fix tests and move ConfigureIndependentProcess functionality to osinstaller
