+++ internal/services/osinstaller/windows-utils.go
@@ -82,3 +82,9 @@ func UnzipOrExtractFiles(installationConfiguration *InstallationConfiguration) e
 
 	return nil
 }
+
+func ConfigureIndependentProcess(cmd *exec.Cmd) {
+	cmd.SysProcAttr = &syscall.SysProcAttr{
+		CreationFlags: syscall.CREATE_NEW_PROCESS_GROUP,
+	}
+}
