--- internal/services/vorpalengine/unix-vorpal-configuration.go
@@ -1,14 +1,0 @@
-//go:build linux || darwin
-
-package vorpalengine
-
-import (
-	"os/exec"
-	"syscall"
-)
-
-func configureIndependentProcess(cmd *exec.Cmd) {
-	cmd.SysProcAttr = &syscall.SysProcAttr{
-		Setsid: true,
-	}
-}
